+++ src/utils/data/assignmentEnum.ts
